+++ ext_template/tasks/locomotion/velocity/config/anymal_d/__init__.py
@@ -7,8 +7,8 @@ from . import agents, flat_env_cfg, rough_env_cfg
 ##
 
 gym.register(
-    id="Template-Velocity-Flat-Anymal-D-v0",
-    entry_point="omni.isaac.orbit.envs:RLTaskEnv",
+    id="Template-Isaac-Velocity-Flat-Anymal-D-v0",
+    entry_point="omni.isaac.lab.envs:ManagerBasedRLEnv",
     disable_env_checker=True,
     kwargs={
         "env_cfg_entry_point": flat_env_cfg.AnymalDFlatEnvCfg,
@@ -17,8 +17,8 @@ gym.register(
 )
 
 gym.register(
-    id="Template-Velocity-Flat-Anymal-D-Play-v0",
-    entry_point="omni.isaac.orbit.envs:RLTaskEnv",
+    id="Template-Isaac-Velocity-Flat-Anymal-D-Play-v0",
+    entry_point="omni.isaac.lab.envs:ManagerBasedRLEnv",
     disable_env_checker=True,
     kwargs={
         "env_cfg_entry_point": flat_env_cfg.AnymalDFlatEnvCfg_PLAY,
@@ -27,8 +27,8 @@ gym.register(
 )
 
 gym.register(
-    id="Template-Velocity-Rough-Anymal-D-v0",
-    entry_point="omni.isaac.orbit.envs:RLTaskEnv",
+    id="Template-Isaac-Velocity-Rough-Anymal-D-v0",
+    entry_point="omni.isaac.lab.envs:ManagerBasedRLEnv",
     disable_env_checker=True,
     kwargs={
         "env_cfg_entry_point": rough_env_cfg.AnymalDRoughEnvCfg,
@@ -37,8 +37,8 @@ gym.register(
 )
 
 gym.register(
-    id="Template-Velocity-Rough-Anymal-D-Play-v0",
-    entry_point="omni.isaac.orbit.envs:RLTaskEnv",
+    id="Template-Isaac-Velocity-Rough-Anymal-D-Play-v0",
+    entry_point="omni.isaac.lab.envs:ManagerBasedRLEnv",
     disable_env_checker=True,
     kwargs={
         "env_cfg_entry_point": rough_env_cfg.AnymalDRoughEnvCfg_PLAY,