+++ docs/CHANGELOG.rst
@@ -7,4 +7,4 @@ Changelog
 Added
 ^^^^^
 
-* Created an initial template for building an extension or project based on Orbit
+* Created an initial template for building an extension or project based on Isaac Lab